\name{pvxmlrpclibs-package}
\alias{pvxmlrpclibs-package}
\alias{pvxmlrpclibs}
\docType{package}
\title{
PathVisio XML-RPC libraries.
}
\description{
Package with the PathVisio XML-RPC libraries, mostly aimed for use
in the pvxmlrpc package. Using the rjava package, the Jena libraries
can also be used directly.
}
}
\author{
Egon Willighagen

Maintainer: Egon Willighagen <egon.willighagen@gmail.com>
}
\keyword{ package }
\keyword{ PathVisio }
\keyword{ pathway }
\keyword{ WikiPathways }
